fix(code-block)!: use script tags
