+++ elements/pfe-code-block/README.md
@@ -1,7 +1,9 @@
 # PatternFly Elements Code Block
-`<pfe-code-block>` is a component that contains two or more lines of ready-only code. The code in a code block can be copied to the clipboard.
+`<pfe-code-block>` is a component that contains two or more lines of ready-only 
+code. The code in a code block can be copied to the clipboard.
 
-Read more about Code Block in the [PatternFly Elements Code Block documentation](https://patternflyelements.org/components/code-block)
+Read more about Code Block in the [PatternFly Elements Code Block 
+documentation](https://patternflyelements.org/components/code-block)
 
 ##  Installation
 
@@ -26,58 +28,50 @@ import '@patternfly/pfe-code-block';
 ## Usage
 ### Basic
 
-Place your code in a `span` with `slot="code"`.
+Place your code in a `script` tag with a [non-javascript mimetype][mime].
+JavaScript snippets **must** use the `text/javascript-sample` script type. 
+Script text content will be automatically dedented.
 
-To add copy-to-clipboard functionality, be sure to import `@patternfly/pfe-clipboard` and add the `pfe-clipboard` to the `action` slot.
 
-<pfe-code-block id="basic">
-  <span slot="code">apiVersion: helm.openshift.io/v1beta1/
-kind: HelmChartRepository
-metadata:
-name: azure-sample-repo
-spec:
-connectionConfig:
-url: https://raw.githubusercontent.com/Azure-Samples/helm-charts/master/docs</span>
-  <pfe-clipboard slot="actions" copy-from="#basic"></pfe-clipboard>
-</pfe-code-block>
+To add copy-to-clipboard functionality, be sure to import 
+`@patternfly/pfe-clipboard` and add the `pfe-clipboard` to the `action` slot.
 
 ```html
-<pfe-code-block id="basic">
-  <span slot="code">apiVersion: helm.openshift.io/v1beta1/
-kind: HelmChartRepository
-metadata:
-name: azure-sample-repo
-spec:
-connectionConfig:
-url: https://raw.githubusercontent.com/Azure-Samples/helm-charts/master/docs</span>
-  <pfe-clipboard slot="actions" copy-from="#basic"></pfe-clipboard>
+<pfe-code-block>
+  <script type="application/openshift">
+    apiVersion: helm.openshift.io/v1beta1/
+    kind: HelmChartRepository
+    metadata:
+    name: azure-sample-repo0oooo00ooo
+    spec:
+    connectionConfig:
+    url: https://raw.githubusercontent.com/Azure-Samples/helm-charts/master/docs
+  </script>
 </pfe-code-block>
 ```
 
 ### Expandable
 
-If a block of code is long and you'd like to hide some of the code to take us less height on the page, place the code you'd like to hide in a `span` with `slot="expandable-code"`. 
+If a block of code is long and you'd like to hide some of the code to take up 
+less height on the page, place the code you'd like to hide in a `script` with 
+`data-expand` attribute present.
 
-It is important that you place the span right next to the `<span slot="code">` because of how the whitespace will be handled. Failing to do this can result in unwanted line breaks. 
-
-<pfe-code-block id="expandable">
-  <span slot="code">apiVersion: helm.openshift.io/v1beta1/
-kind: HelmChartRepository
-metadata:
-name: azure-sample-repo</span><span slot="expandable-code">spec:
-connectionConfig:
-url: https://raw.githubusercontent.com/Azure-Samples/helm-charts/master/docs</span>
-  <pfe-clipboard slot="actions" copy-from="#expandable"></pfe-clipboard>
-</pfe-code-block>
+It is important that you place the span right next to the first `script`, 
+because of how the whitespace will be handled. Failing to do this can result in 
+unwanted line breaks. 
 
 ```html
-<pfe-code-block id="expandable">
-  <span slot="code">apiVersion: helm.openshift.io/v1beta1/
-kind: HelmChartRepository
-metadata:
-name: azure-sample-repo</span><span slot="expandable-code">spec:
-connectionConfig:
-url: https://raw.githubusercontent.com/Azure-Samples/helm-charts/master/docs</span>
-  <pfe-clipboard slot="actions" copy-from="#expandable"></pfe-clipboard>
+<pfe-code-block id="expandable-code">
+  <script type="application/openshift">
+    apiVersion: helm.openshift.io/v1beta1/
+    kind: HelmChartRepository
+    metadata:
+    name: azure-sample-repo</script><script type="application/openshift" 
+    data-expand>
+    spec:
+    connectionConfig:
+    url: https://raw.githubusercontent.com/Azure-Samples/helm-charts/master/docs
+  </script>
+  <pfe-clipboard slot="actions" copy-from="#expandable-code"></pfe-clipboard>
 </pfe-code-block>
 ```
