+++ elements/pfe-code-block/demo/pfe-code-block.js
@@ -1,6 +1,2 @@
 import '@patternfly/pfe-code-block';
 import '@patternfly/pfe-clipboard';
-
-const root = document.querySelector('[data-demo="pfe-code-block"]')?.shadowRoot ?? document;
-
-root.querySelector('pfe-code-block');
