+++ elements/pfe-code-block/docs/pfe-code-block.md
@@ -2,13 +2,15 @@
   Code Block is a component that contains two or more lines of read-only code. The code in a code block can be copied to the clipboard.
 
   <pfe-code-block id="code">
-    <span slot="code">apiVersion: helm.openshift.io/v1beta1/
-kind: HelmChartRepository
-metadata:
-name: azure-sample-repo
-spec:
-connectionConfig:
-url: https://raw.githubusercontent.com/Azure-Samples/helm-charts/master/docs</span>
+    <script type="application/openshift">
+      apiVersion: helm.openshift.io/v1beta1/
+      kind: HelmChartRepository
+      metadata:
+      name: azure-sample-repo
+      spec:
+      connectionConfig:
+      url: https://raw.githubusercontent.com/Azure-Samples/helm-charts/master/docs
+    </script>
     <pfe-clipboard slot="actions" copy-from="#code"></pfe-clipboard>
   </pfe-code-block>
 {% endrenderOverview %}
@@ -16,30 +18,36 @@ url: https://raw.githubusercontent.com/Azure-Samples/helm-charts/master/docs</sp
 {% band header="Usage" %}
 ### Basic
 
-Place your code in a `span` with `slot="code"`.
+Place your code in a `script` tag with a [non-javascript mimetype][mime].
+JavaScript snippets **must** use the `text/javascript-sample` script type. 
+Script text content will be automatically dedented.
 
 To add copy-to-clipboard functionality, be sure to import `@patternfly/pfe-clipboard` and add the `pfe-clipboard` to the `action` slot.
 
 <pfe-code-block id="basic">
-  <span slot="code">apiVersion: helm.openshift.io/v1beta1/
-kind: HelmChartRepository
-metadata:
-name: azure-sample-repo
-spec:
-connectionConfig:
-url: https://raw.githubusercontent.com/Azure-Samples/helm-charts/master/docs</span>
+  <script type="application/openshift">
+    apiVersion: helm.openshift.io/v1beta1/
+    kind: HelmChartRepository
+    metadata:
+    name: azure-sample-repo
+    spec:
+    connectionConfig:
+    url: https://raw.githubusercontent.com/Azure-Samples/helm-charts/master/docs
+  </script>
   <pfe-clipboard slot="actions" copy-from="#basic"></pfe-clipboard>
 </pfe-code-block>
 
 ```html
 <pfe-code-block id="basic">
-  <span slot="code">apiVersion: helm.openshift.io/v1beta1/
-kind: HelmChartRepository
-metadata:
-name: azure-sample-repo
-spec:
-connectionConfig:
-url: https://raw.githubusercontent.com/Azure-Samples/helm-charts/master/docs</span>
+  <script type="application/openshift">
+    apiVersion: helm.openshift.io/v1beta1/
+    kind: HelmChartRepository
+    metadata:
+    name: azure-sample-repo
+    spec:
+    connectionConfig:
+    url: https://raw.githubusercontent.com/Azure-Samples/helm-charts/master/docs
+  </script>
   <pfe-clipboard slot="actions" copy-from="#basic"></pfe-clipboard>
 </pfe-code-block>
 ```
@@ -48,27 +56,35 @@ url: https://raw.githubusercontent.com/Azure-Samples/helm-charts/master/docs</sp
 
 If a block of code is long and you'd like to hide some of the code to take us less height on the page, place the code you'd like to hide in a `span` with `slot="expandable-code"`. 
 
-It is important that you place the span right next to the `<span slot="code">` because of how the whitespace will be handled. Failing to do this can result in unwanted line breaks. 
+It is important that you place the span right next to the `<script type="application/openshift">` because of how 
+the whitespace will be handled. Failing to do this can result in unwanted line 
+breaks. 
 
-<pfe-code-block id="expandable">
-  <span slot="code">apiVersion: helm.openshift.io/v1beta1/
-kind: HelmChartRepository
-metadata:
-name: azure-sample-repo</span><span slot="expandable-code">spec:
-connectionConfig:
-url: https://raw.githubusercontent.com/Azure-Samples/helm-charts/master/docs</span>
-  <pfe-clipboard slot="actions" copy-from="#expandable"></pfe-clipboard>
+<pfe-code-block id="expandable-code">
+  <script type="application/openshift">
+    apiVersion: helm.openshift.io/v1beta1/
+    kind: HelmChartRepository
+    metadata:
+    name: azure-sample-repo</script><script type="application/openshift" data-expand>
+    spec:
+    connectionConfig:
+    url: https://raw.githubusercontent.com/Azure-Samples/helm-charts/master/docs
+  </script>
+  <pfe-clipboard slot="actions" copy-from="#expandable-code"></pfe-clipboard>
 </pfe-code-block>
 
 ```html
-<pfe-code-block id="expandable">
-  <span slot="code">apiVersion: helm.openshift.io/v1beta1/
-kind: HelmChartRepository
-metadata:
-name: azure-sample-repo</span><span slot="expandable-code">spec:
-connectionConfig:
-url: https://raw.githubusercontent.com/Azure-Samples/helm-charts/master/docs</span>
-  <pfe-clipboard slot="actions" copy-from="#expandable"></pfe-clipboard>
+<pfe-code-block id="expandable-code">
+  <script type="application/openshift">
+    apiVersion: helm.openshift.io/v1beta1/
+    kind: HelmChartRepository
+    metadata:
+    name: azure-sample-repo</script><script type="application/openshift" data-expand>
+    spec:
+    connectionConfig:
+    url: https://raw.githubusercontent.com/Azure-Samples/helm-charts/master/docs
+  </script>
+  <pfe-clipboard slot="actions" copy-from="#expandable-code"></pfe-clipboard>
 </pfe-code-block>
 ```
 {% endband %}
@@ -84,3 +100,5 @@ url: https://raw.githubusercontent.com/Azure-Samples/helm-charts/master/docs</sp
 {% renderCssCustomProperties %}{% endrenderCssCustomProperties %}
 
 {% renderCssParts %}{% endrenderCssParts %}
+
+[mime]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/MIME_types#textjavascript
